+++ CHANGES.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.5 (dev)
+------------------
+
+- Added '--level' option to 'doorstop add' to force an Item level
+- Added warnings for duplicate Item levels in a Document
+
 0.4.3 (2014/03/18)
 ------------------